Create guess.py
When you run this script, it will generate a random number between 1 and 100, and then ask the user to guess what it is. If the user's guess is too low or too high, the script will give them a hint and ask them to guess again. Once the user guesses the correct number, the script will congratulate them and end.